refactor(ui): extract panel interaction handling
